jmbo_calendar/tests/requirements.txt
psycopg2
django-tastypie<0.10
